test(codec): move malformed bounded string regression
