+++ commons/zenoh-codec/tests/codec.rs
@@ -271,19 +271,6 @@ fn codec_string_bounded() {
 }
 
 #[test]
-fn codec_bounded_string_rejects_declared_length_larger_than_input() {
-    // This is a length-prefixed string with a declared length of 2, but only 1
-    // byte follows. The bounded byte/string decoder must reject the truncated
-    // payload instead of trusting the declared length for allocation.
-    let bytes = [2, b'a'];
-
-    let codec = Zenoh080Bounded::<usize>::new();
-    let mut reader = bytes.as_slice().reader();
-    let decoded: Result<String, _> = codec.read(&mut reader);
-    assert!(decoded.is_err());
-}
-
-#[test]
 fn codec_zid() {
     run!(ZenohIdProto, ZenohIdProto::default());
 }
